starlark: refactor how target builds are evaluated
Previously, logic for building a target was defined by the presence of
a `BuildTarget` trait. We had to cast a `Value` to a concrete type then
invoke its `.build()` via that trait.

Unfortunately, due to the way types are represented in `Value`, I don't
think it was possible to dynamically cast to any type that implements
`BuildTarget` because we don't control the constraints on `Value` type
conformance. So we had to match the type name and attempt a cast to its
concrete type before calling `.build()`. This approach wasn't generic
and made defining buildable types in different crates difficult.

This commit refactors how target building is performed. Essentially, we
replace the `BuildTarget` trait with a `build()` method implemented in
Starlark. The code for building a `Value` simply looks for and calls the
`build()` method of the `Value`. Then the type itself is responsible
for instantiating a context used to influence the build.

We can possibly get rid of the highly dynamic `BuildContext` type now
and rely on each crate/type to reference their own context. I'll have
to think more about this.